backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose==3.3.0
ecdsa>=0.18.0
rsa>=4.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
numpy==1.24.4
pandas==2.0.3
pyarrow>=14.0.0
openpyxl==3.1.2
scikit-learn==1.5.0
python-dotenv==1.0.1
requests>=2.31.0
reportlab==4.2.0
python-docx==1.1.0
psycopg2-binary
